get_labels_stats/requirements.txt
pyorthanc
python-dotenv